==> Code/01_Combine_Function.ipynb <==
### The purpose of this note book is to combine 2019 and 2020 MLB stats for both pitchers and batters.  Also clean testing data to match training data for predictions.

#imports
import pandas as pd

# Batter Functions

    # Create function for batters
def batter_combine(bat_19, bat_20):
        # step 1 merge
        batters = bat_19.merge(bat_20, how='left', on='Name')

        # step 2 handle nulls
        batters.fillna(0, inplace=True)

        # step 3 create list for integer stats
        # create a list of integer stats
        col_list = ['GMS_', 'AB_', 'R_', 'H_', '2B_', '3B_', 'HR_', 'RBI_', 'BB_', 'SO_', 'SB_', 'CS_', 'PTS_']

        # for each column in the list of int stats
        for col in col_list:
            # create name for col
            col_name = str(col).replace('_','')
            # create list for data
            col_data = []
            # set counter for y column
            counter = 0

            # for each row in the specific column
            for row in batters[f'{col}x']:
                # sum similar cols somehow; possibly use x,y f string
                col_data.append(row + batters[f'{col}y'][counter])
                # increase counter
                counter += 1

            # add col to df
            batters[col_name] = col_data
            # drop x, y cols for col
            batters.drop(columns=[f'{col}x', f'{col}y'], inplace=True)

        # step 4 drop and rename columns
        # create drop list
        drop_list = ['Rank_', 'AVG_', 'OBP_', 'SLG_', 'OPS_']
        # loop through list and drop 
        for col in drop_list:
            batters.drop(columns=[f'{col}x', f'{col}y'], inplace=True)
        # drop duplicate cols
        batters.drop(columns=['Team_y', 'POS_y'], inplace=True)
        # rename remaining team, pos columns
        batters.rename(columns={'Team_x':'Team', 'POS_x': 'POS'}, inplace=True)

        # step 5 total bases column
        # total bases
        total_bases = []
        counter2 = 0
        for row in batters['H']:
            # define singles
            single = row - (batters['2B'][counter2] + batters['3B'][counter2] + batters['HR'][counter2])
            # create total bases for each player and add to total bases list 
            total_bases.append(single + ((batters['2B'][counter2])*2) + ((batters['3B'][counter2])*3) + ((batters['HR'][counter2])*4))
            # increase counter
            counter2 += 1

        # add list as column to dataframe
        batters['TB'] = total_bases

        # step 6 create percentage stat categories
        # create batting avg list
        bat_avg = []
        # create slugging percentage list 
        slg_list = []
        # create on base percentage list
        obp = []
        # create ops list
        ops = []
        # fantasy points per game
        fppg = []
        # for each row calculate stats
        for i in batters.index:
            #calculate batting average and add to list 
            bat_avg.append(round((batters['H'][i]/batters['AB'][i]), 4))
            #calculate slugging percentage and add to list
            slg_list.append(round((batters['TB'][i]/batters['AB'][i]), 4))
            #calculate on base percentage and add to list 
            obp.append(round(((batters['H'][i]+batters['BB'][i])/(batters['AB'][i]+batters['BB'][i])),4))
            #calculate fppg and add to list
            fppg.append(batters['PTS'][i]/batters['GMS'][i])

        # create seperate loop for ops
        for x,y in zip(obp, slg_list):
            ops.append(round((x+y), 4))

        # add stats to dataframe
        batters['AVG'] = bat_avg
        batters['SLG'] = slg_list
        batters['OBP'] = obp
        batters['OPS'] = ops
        batters['FPPG'] = fppg
    
        # step 7 add Iso to df
        # create list for ISO
        iso = []
        # for each row calculate ISO
        for i in batters.index:
            # calculate ISO
            iso.append(round((batters['SLG'][i]-batters['AVG'][i]), 4))
        # add iso to data frame
        batters['ISO'] = iso
        
        # step 8 drop nulls if any 
        batters.dropna(inplace=True)
        
        # step 10 clean Pos column
        pos_list = []
        for pos in batters['POS']:
            if pos == 'LF' or pos == 'CF' or pos == 'RF' or pos == 'DH':
                pos_list.append('OF')
            else:
                pos_list.append(pos)
        # replace pos with list
        batters['POS'] = pos_list
        
        # step 9 reorder columns
        batters = batters[['Name', 'Team', 'POS', 'GMS', 'AB', 'R', 'H', '2B', '3B', 'HR', 'RBI',
                           'BB', 'SO', 'SB', 'CS', 'TB', 'AVG', 'OBP', 'SLG', 'OPS', 'ISO', 'PTS',
                           'FPPG']]

        # RETURN
        return batters

# seperate function for test data 
def testing_data_clean(batters):
        # total bases
        total_bases = []
        counter2 = 0
        for row in batters['H']:
            # define singles
            single = row - (batters['2B'][counter2] + batters['3B'][counter2] + batters['HR'][counter2])
            # create total bases for each player and add to total bases list 
            total_bases.append(single + ((batters['2B'][counter2])*2) + ((batters['3B'][counter2])*3) + ((batters['HR'][counter2])*4))
            # increase counter
            counter2 += 1

        # add list as column to dataframe
        batters['TB'] = total_bases
        
        # create list for ISO
        iso = []
        # fantasy points per game
        fppg = []
        
        for i in batters.index:
            # calculate ISO
            iso.append(round((batters['SLG'][i]-batters['AVG'][i]), 4))
            #calculate fppg and add to list
            fppg.append(batters['PTS'][i]/batters['GMS'][i])
        # add to df
        batters['ISO'] = iso
        batters['FPPG'] = fppg
        
        # drop rank
        batters.drop(columns='Rank', inplace=True)
        
        # clean Pos column
        pos_list = []
        for pos in batters['Pos']:
            if pos == 'LF' or pos == 'CF' or pos == 'RF' or pos == 'DH':
                pos_list.append('OF')
            else:
                pos_list.append(pos)
        # replace pos with list
        batters['Pos'] = pos_list
        
        # match same column order as train data
        batters = batters[['Name', 'Team', 'Pos', 'GMS', 'AB', 'R', 'H', '2B', '3B', 'HR', 'RBI',
                           'BB', 'SO', 'SB', 'CS', 'TB', 'AVG', 'OBP', 'SLG', 'OPS', 'ISO', 'PTS',
                           'FPPG']]
        return batters

---

### Importing, cleaning and saving batter testing data.

# read in test data
batters = pd.read_csv('../Collection_data/FantasyData/fanduel_2021C - batters.csv')

batters = testing_data_clean(batters)

# export to folder 
batters.to_csv('../Modeling_Data/batter_test_data_c.csv', index=False)

### Importing, cleaning and saving batter training data.

# read in 2019 and 2020 batter data
bat_19 = pd.read_csv('../Collection_data//FantasyData/fanduel_2019 - batters.csv')
bat_20 = pd.read_csv('../Collection_data//FantasyData/fanduel_2020 - batters.csv')

# test function on past data
batter = batter_combine(bat_19, bat_20)

batter.to_csv('../Modeling_Data/batter_data_modeling.csv', index=False)

---

# Pitcher Functions

def combine_pitchers(df1, df2):
    # Step 1 merge dataframes
    pitchers = df1.merge(df2, how='left', on='Name')

    # Step 2 fill nulls with zero
    pitchers.fillna(0, inplace=True)

    # step 3 create column list for integer stats to be summed and add to dataframe, also delete old columns
    col_list = ['W_', 'L_', 'GMS_', 'GS_', 'SV_', 'IP_', 'H_', 'R_', 'ER_', 'HR_', 'BB_', 'SO_', 'PTS_']

    # for each column in the list of int stats
    for col in col_list:
        # create name for col
        col_name = str(col).replace('_','')
        # create list for data
        col_data = []
        # set counter for y column
        counter = 0

        # for each row in the specific column
        for row in pitchers[f'{col}x']:
            # sum similar cols somehow; possibly use x,y f string
            col_data.append(row + pitchers[f'{col}y'][counter])
            # increase counter
            counter += 1

        # add col to df
        pitchers[col_name] = col_data
        # drop x, y cols for col
        pitchers.drop(columns=[f'{col}x', f'{col}y'], inplace=True)
       
    # step 4 create avg column for opponent batting average
    # create avg list
    avg_list = []
    # reset counter
    counter = 0
    # for each row in avg calculate average
    for row in pitchers['AVG_x']:
        # calculate average of AVG
        avg_list.append(round(((row + pitchers['AVG_y'][counter])/2), 4))
    
    # add col to df
    pitchers['AVG'] = avg_list
    # drop x, y cols for col
    pitchers.drop(columns=['AVG_x', 'AVG_y'], inplace=True)

    # step 5 drop and rename columns
    pitchers.drop(columns=['Team_y', 'POS_y'], inplace=True)
    # rename remaining team, pos columns
    pitchers.rename(columns={'Team_x':'Team', 'POS_x': 'POS'}, inplace=True)
    # create drop list
    drop_list = ['Rank_', 'ERA_', 'WHIP_']
    # loop through list and drop 
    for col in drop_list:
        pitchers.drop(columns=[f'{col}x', f'{col}y'], inplace=True)

    # step 6 Add percentage stats to pitchers. ERA and WHIP.
    # create era list
    era_list = []
    # create whip list 
    whip_list = []
    # fantasy points per game
    fppg = []
    # for each row calculate stats
    for i in pitchers.index:
        #calculate batting average and add to list 
        era_list.append(round(((9*pitchers['ER'][i])/pitchers['IP'][i]), 4))
        #calculate slugging percentage and add to list
        whip_list.append(round((pitchers['BB'][i] + pitchers['H'][i]/pitchers['IP'][i]), 4))
        #calculate fppg and add to list
        fppg.append(pitchers['PTS'][i]/pitchers['GMS'][i])

    # add stats to dataframe
    pitchers['ERA'] = era_list
    pitchers['WHIP'] = whip_list
    pitchers['FPPG'] = fppg

    # step 7 drop nulls and any pitcher with less than 0 fantasy point per game
    pitchers.dropna(inplace=True)
    pitchers = pitchers.loc[pitchers['PTS']>0]
    
    # step 8 clean Pos column
    pos_list = []
    for pos in pitchers['POS']:
        if pos == 'P':
            pos_list.append('SP')
        else:
            pos_list.append(pos)
        # replace pos with list
    pitchers['POS'] = pos_list

    # return
    return pitchers

# testing data function
def testing_pitcher_data_clean(pitchers):
    
    # step 1 drop columns
    pitchers.drop(columns='Rank')
    
    # step 2 add FPPG
    fppg = []
    # for each row calculate stats
    for i in pitchers.index:
        #calculate fppg and add to list
        fppg.append(pitchers['PTS'][i]/pitchers['GMS'][i])
    # add stats to dataframe
    pitchers['FPPG'] = fppg
    
    # step 3 reorder columns to match train
    pitchers = pitchers[['Name', 'Team', 'Pos', 'W', 'L', 'GMS', 'GS', 'SV', 'IP', 'H', 'R',
       'ER', 'HR', 'BB', 'SO', 'PTS', 'ERA', 'AVG', 'WHIP', 'FPPG']]
    
    # step 4 clean Pos column
    pos_list = []
    for pos in pitchers['Pos']:
        if pos == 'P':
            pos_list.append('SP')
        else:
            pos_list.append(pos)
        # replace pos with list
    pitchers['Pos'] = pos_list
    
    # step 5 drop any pitcher with less than 0 fantasy points per game
    pitchers = pitchers.loc[pitchers['PTS']>0]
    
    # return
    return pitchers

### Importing, cleaning and saving pitcher training data.

# test pitcher function
pitch_19 = pd.read_csv('../Collection_data/FantasyData/fanduel_2019 - pitchers.csv')
pitch_20 = pd.read_csv('../Collection_data/FantasyData//fanduel_2020 - pitchers.csv')

# call function
pitchers = combine_pitchers(pitch_19, pitch_20)

# export to csv
pitchers.to_csv('../Modeling_Data/pitcher_data_modeling.csv', index=False)

### Importing, cleaning and saving pitcher testing data.

# read in test data for pitchers
pitchers_test = pd.read_csv('../Collection_data//FantasyData/fanduel_2021C - pitchers.csv')

# clean pitcher test data with function
pitchers_test = testing_pitcher_data_clean(pitchers_test)

# export to csv
pitchers_test.to_csv('../Modeling_Data/pitcher_test_data_c.csv', index=False)
