==> Code/03_GLM.ipynb <==
Will use a generalized linear model in this notebook.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import statsmodels.api as sm

from sklearn.preprocessing import StandardScaler

# Batter

# read in batter training data
batter_train = pd.read_csv('../Modeling_Data/batter_data_modeling.csv')
# read in batter testing data
batter_test = pd.read_csv('../Modeling_Data/batter_test_data.csv')

# set up X and y from training data
X = batter_train.drop(columns=['Name', 'Team', 'POS', 'PTS', 'FPPG'])
y = batter_train['FPPG']

# add constant to X
X = sm.add_constant(X)

# scale data
sc = StandardScaler()
# fit and transform X
X_sc = sc.fit_transform(X)

# GLM
glm_batter = sm.GLM(y, X_sc, family=sm.families.Poisson(link=sm.families.links.log())).fit()

results_summary = glm_batter.summary()

glm_batter.summary()

# create data frame of coefficients
# save as html
results_as_html = results_summary.tables[1].as_html()
# read in html
results_df = pd.read_html(results_as_html, header=0, index_col=0)[0]
# set index from X columns
results_df['Features'] = X.columns
results_df.set_index('Features', inplace=True)

# sort values
results_df.sort_values(by='coef')

# Interpert Coefficient Values

---

# Pitcher

# read in pitcher training data
pitcher_train = pd.read_csv('../Modeling_Data/pitcher_data_modeling.csv')
# read in batter testing data
pitcher_test = pd.read_csv('../Modeling_Data/pitcher_test_data.csv')

# set up X and y from training data
X = pitcher_train.drop(columns=['Name', 'Team', 'POS', 'PTS', 'AVG', 'FPPG'])
y = pitcher_train['FPPG']

# add constant
X = sm.add_constant(X)

# scale data
sc = StandardScaler()
# fit and transform X
X_sc = sc.fit_transform(X)

# GLM
glm_pitcher = sm.GLM(y, X_sc, family=sm.families.Poisson(link=sm.families.links.log())).fit()

results_summary = glm_pitcher.summary()

glm_pitcher.summary()

# create data frame of coefficients
# save as html
results_as_html = results_summary.tables[1].as_html()
# read in html
results_df = pd.read_html(results_as_html, header=0, index_col=0)[0]
# set index from X columns
results_df['Features'] = X.columns
results_df.set_index('Features', inplace=True)

# sort values
results_df.sort_values(by='coef')

# Interpert

The coefficients provided in the GLM are very useful in this project and can be used provide insights in to the strategy of building a solid line up.
