==> Code/04_Yahoo_PlayerList_Function.ipynb <==
# imports
import pandas as pd

# files
# yahoo 
yahoo = pd.read_csv('../Collection_data/Yahoo/Yahoo_DF_player_export.csv')
# projections
pitcher_proj = pd.read_csv('../Projections/pitcher_projections_2021.csv')
batter_proj = pd.read_csv('../Projections/batter_projections_2021.csv')

yahoo.head(1)

pitcher_proj.head(1)

## Steps to merge projections with yahoo player list
1. create name column 
2. drop unneeded columns
### starting is a very interesting column for position players if populated
3. reorder columns
4. create filter for injury column, as healthy players have ' '
5. split dataframe batters and pitchers 

# yahoo naming function
def yahoo_name(yahoo):
    # create empty list for names
    player_list = []
    # for each row concat names and append to list 
    for row in yahoo.index:
        player_list.append(yahoo['First Name'][row] + ' ' + yahoo['Last Name'][row])
        # make new column for name
    yahoo['Name'] = player_list
    return yahoo

# step 1
yahoo = yahoo_name(yahoo)

# step 2
yahoo.drop(columns=['First Name', 'Last Name', 'Game', 'Time', 'Starting'], inplace=True)

# step 3 
yahoo = yahoo[['ID', 'Name', 'Position', 'Team', 'Opponent', 'Salary', 'FPPG', 'Probable Pitcher', 'Injury Status']].copy()

# step 4
yahoo = yahoo.loc[yahoo['Injury Status']==' '].copy()

# step 5
batters = yahoo.loc[yahoo['Position']!='P']
pitchers = yahoo.loc[yahoo['Position']=='P']

## Pticher steps
1. filter by only starting pitchers
2. read in projections and merge 
3. drop any nulls
4. reorder columns
5. rename columns

# step 1
pitchers = pitchers.loc[pitchers['Probable Pitcher']=='Yes']

# step 2
pitcher_projections = pitchers.merge(pitcher_proj, how='left', on='Name')

# step 3
pitcher_projections.dropna(inplace=True)

# step 4
pitcher_projections = pitcher_projections[['ID', 'Position', 'Name', 'Salary', 'Team_x', 'Opponent', 'Projected_FPPG']].copy()

# step 5
pitcher_projections.rename(columns={'Team_x':'Team'}, inplace=True)

pitcher_projections

## Batter Steps
1. merge
2. drop nulls
3. reorder
4. rename

batter_projections = batters.merge(batter_proj, how='left', on='Name')

batter_projections.dropna(inplace=True)

batter_projections = batter_projections[['ID', 'Position', 'Name', 'Salary', 'Team_x', 'Opponent','Projected_FPPG']].copy()

batter_projections.rename(columns={'Team_x':'Team'}, inplace=True)

batter_projections

# following are steps for creating a line up
# set cap for yahoo
salary_cap = 200

# sort pitcher by fppg projections
pitcher_projections.sort_values(by='Projected_FPPG', ascending=False, inplace=True, ignore_index=True)

# create a player list to 
lineup = []
# add top two pitchers based on projections
lineup.append(pitcher_projections.values[0])
lineup.append(pitcher_projections.values[1])

# need to update remaining salary
salary_cap -= (pitcher_projections['Salary'][0]+pitcher_projections['Salary'][1])


# with updated salary fill remaining roster based on position and highest fppg
# create position list for remaining roster spots
position_list = ['C', '1B', '2B', '3B', 'SS', 'OF', 'OF', 'OF']
# sort batters by FPPG
batter_projections.sort_values(by='Projected_FPPG', ascending=False, inplace=True, ignore_index=True)

# create count based on remaining positions
sal_count = 8

# create average salary variable for remaining players
avg_sal = salary_cap/sal_count

# create for loop for each position in list to take highest fppg
for pos in position_list:
    # setting counter to increase if player is already in list
    # this is inside the for loop beacuse it needs to be per position
    counter = 0
    # if salary greater than average move to next player
    for salary in batter_projections.loc[batter_projections['Position'] == pos]['Salary']:
        # test if salary is greater than average if it is increase counter
        if batter_projections.loc[batter_projections['Position'] == pos]['Salary'].values[counter] > avg_sal or batter_projections.loc[batter_projections['Position'] == pos]['Salary'].values[counter] > salary_cap:
            counter += 1
        else:
            # if less than average add player to list
            lineup.append(batter_projections.loc[batter_projections['Position'] == pos].values[counter])
            # drop player so no duplicates are added
            batter_projections.drop(batter_projections.loc[batter_projections['Position'] == pos].index.values[counter], inplace=True)
            # decrease sal_count
            sal_count -= 1
            # decrease salary cap
            salary_cap -= batter_projections.loc[batter_projections['Position'] == pos]['Salary'].values[counter]
            # create new average salary
            avg_sal = salary_cap/sal_count
            break
# create dataframe of lineup
df_lineup = pd.DataFrame(lineup, columns=['Id', 'Position', 'Name', 'Salary', 'Team', 'Opponent','Projected_FPPG'])

df_lineup

df_lineup['Salary'].sum()
